requirements.txt
altair==5.5.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
catboost==1.2.8
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cryptography-fernet==0.1.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.4
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
folium==0.20.0
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.21
idna==3.10
imbalanced-learn==0.12.2
imblearn==0.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lightgbm==4.6.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
missingno==0.5.2
mistune==3.1.3
narwhals==1.39.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.1
plotly-express==0.4.1
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.4.2
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.45.1
streamlit-option-menu==0.4.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.27.1
virtualenvwrapper-win==1.2.7
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xgboost==3.0.2
xyzservices==2025.4.0
